requirements.txt
Django==2.1.4
django-cors-middleware==1.3.1
django-extensions==2.1.4
django-rest-swagger==2.1.0
djangorestframework==3.9.0
psycopg2-binary==2.7.6.1
pycodestyle==2.4.0
dj-database-url==0.5.0
whitenoise==4.1.2
gunicorn==19.9.0
django-heroku==0.3.1
PyJWT==1.7.1
<<<<<<< HEAD
pylint==2.2.2
pytz==2018.9
cloudinary==1.15.0
Pillow==5.4.1
django-taggit==0.23.0
requests==2.21.0
simplejson==3.16.0
=======
>>>>>>> feat (Profile) user can create a profile
six==1.12.0
pytz==2018.9
cloudinary==1.15.0
Pillow==5.4.1
